Updated Textbox Layouts

+++ Assets/Scripts/Inventory/ItemScript.cs
@@ -22,7 +22,7 @@ public class ItemScript : ScriptableObject
     [Header("Egg Interaction")]
     public bool isEggItem = false; 
     public GameObject eggPlacementPrefab;
-    public int sellValue;
+    public float sellValue;
